requirements.txt
asttokens==2.2.1
backcall==0.2.0
colorama==0.4.6
comm==0.1.2
comtypes==1.1.14
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
ipykernel==6.21.1
ipython==8.9.0
jedi==0.18.2
jupyter_client==8.0.2
jupyter_core==5.2.0
matplotlib-inline==0.1.6
MouseInfo==0.1.3
nest-asyncio==1.5.6
numpy==1.24.2
opencv-python==4.7.0.68
packaging==23.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
prompt-toolkit==3.0.36
psutil==5.9.4
pure-eval==0.2.2
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
Pygments==2.14.0
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
python-dotenv==0.21.1
pytweening==1.0.4
pywin32==305
pywinauto==0.6.8
pyzmq==25.0.0
six==1.16.0
stack-data==0.6.2
tornado==6.2
traitlets==5.9.0
wcwidth==0.2.6
